+++ CHANGELOG.md
@@ -224,3 +224,4 @@
 * Added method to get list of supported DEX's
 * Added method to change DEX's
 * Added method to get list of supported tokens
+* Added method to get exchange rates